requirements.txt
openai
chromadb
pydantic
fastapi
requests
beautifulsoup4
python-dotenv